fix address api

+++ apps/order-service/src/order/order.service.ts
@@ -1275,6 +1275,10 @@ export class OrderService implements IOrderService {
       );
     }
 
+    if (itemIds.length === 0) {
+      return;
+    }
+
     await this.orderRepository.deleteCartItems(itemIds);
   }
